+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Return value of `NodeHistoryRepository->getByPath()` changed from `Node` to `NodeHistory`
+
 ## [1.0.6] - 2023-04-13
 
 ### Fixed